Refactor tests and routes for improved authorization and error handling
- Updated ServerCrudPatternConsistencyTest to reflect new middleware permissions and route changes.
- Modified ServerReleasesUpdatedNotificationTest and ServerScriptsUpdatedNotificationTest to check for updated job dispatching syntax.
- Enhanced ServerServiceManualSshTest to include additional parameters in secret setting.
- Adjusted WebsiteCrudFlowPatternTest and WebsiteCrudPatternConsistencyTest for route and controller assertions.
- Refined WebsiteDeleteCleanupFlowTest to ensure correct dispatching of delete jobs.
- Updated DashboardTest to reflect changes in navigation structure and permissions.
- Added test routes for authorization checks in testing environment.
- Removed deprecated theme configuration file and adjusted related tests.
- Improved error handling tests to ensure proper rendering of error pages.
- Updated various user management tests to streamline assertions and improve clarity.

+++ modules/CMS/tests/Feature/PostCrudMigrationTest.php
@@ -169,7 +169,7 @@ class PostCrudMigrationTest extends TestCase
             ->assertInertia(fn (Assert $page): Assert => $page
                 ->component('cms/posts/edit')
                 ->where('pickerFilters.picker', '1')
-                ->where('uploadSettings.max_files_per_upload', config('media.max_files_per_upload', 10))
+                ->where('uploadSettings.max_files_per_upload', (int) config('media.max_files_per_upload', 10))
                 ->where('uploadSettings.upload_route', route('app.media.upload-media'))
             );
     }
